OSIS-2255 Move simplified volume to edition volume
