+++ base/tests/forms/test_learning_unit_partim.py
@@ -29,9 +29,8 @@ import factory
 import factory.fuzzy
 from django.forms import model_to_dict
 from django.test import TestCase
-from django.utils.translation import ugettext_lazy as _
 
-from base.forms.learning_unit.simplified_volume import SimplifiedVolumeManagementForm
+from base.forms.learning_unit.edition_volume import SimplifiedVolumeManagementForm
 from base.forms.learning_unit.entity_form import EntityContainerBaseForm
 from base.forms.learning_unit.learning_unit_create import LearningUnitYearModelForm, \
     LearningUnitModelForm, LearningContainerYearModelForm, LearningContainerModelForm
